+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Changelog
 0.8.0 (2020-12-05)
 -------------------
 - Serializing ``dict`` keys that are not strings (can be overridden).
+- Multiple bug fixes regarding mappings that are not ``dict``.
 
 0.7.0 (2020-10-07)
 -------------------
@@ -10,7 +11,7 @@ Changelog
 
 0.6.1 (2020-09-06)
 -------------------
-- Hack to support deserialization with `obj_type` that is a generic type.
+- Hack to support deserialization with ``obj_type`` that is a generic type.
 
 0.6.0 (2020-09-06)
 -------------------
@@ -22,5 +23,5 @@ Changelog
 
 0.4.0 (2020-05-08)
 -------------------
-- Supporting preservation of iterable types even when `type_key` is `None`, through type hints.
-- Changed all warnings to use `warnings` instead of `logging`.
+- Supporting preservation of iterable types even when ``type_key`` is ``None``, through type hints.
+- Changed all warnings to use ``warnings`` instead of ``logging``.